LPS-159329 Control what we show for each action, since we want to show the control menu entry for display pages and page templates too, not only content pages
